+++ Kickstarter-iOS/Views/Cells/Project Page Campaign Cells/ImageViewElementCell.swift
