+++ Kickstarter-iOS/Views/Cells/Project Page Campaign Cells/VideoViewElementCell.swift
