+++ Library/Project Page Campaign HTML Parser/Element+Helpers.swift
@@ -1,0 +1,108 @@
+import SwiftSoup
+
+extension Element {
+  func extractViewElementTypeFromDiv() -> ViewElementType? {
+    var type: ViewElementType?
+
+    let emptyChildren = children().isEmpty()
+    let emptyChildrensChildren = !emptyChildren ? children()[0].children().isEmpty() : true
+
+    if self.isImageStructure(),
+      !emptyChildrensChildren,
+      children()[0].children()[0].tag().getName() == ViewElementType.image.rawValue {
+      type = .image
+    } else if self.isIframeStructure(),
+      !emptyChildrensChildren,
+      children()[0].tag().getName() == ViewElementType.externalSources.rawValue {
+      type = .externalSources
+    }
+
+    return type
+  }
+
+  func isIframeStructure() -> Bool {
+    let templateDivAttributes = getAttributes()?.filter { attribute in
+      let classKey = attribute.getKey() == HTMLRawText.Base.htmlClass.rawValue
+      let templateOembedValue = attribute.getValue() == HTMLRawText.KSRSpecific.templateOembed.rawValue
+
+      return classKey && templateOembedValue
+    }
+
+    if let templateDivAttributesValue = templateDivAttributes {
+      return !templateDivAttributesValue.isEmpty
+    }
+
+    return false
+  }
+
+  func isImageStructure() -> Bool {
+    let templateDivAttributes = getAttributes()?.filter { attribute in
+      let classKey = attribute.getKey() == HTMLRawText.Base.htmlClass.rawValue
+      let templateAssetValue = attribute.getValue() == HTMLRawText.KSRSpecific.templateAsset.rawValue
+
+      return classKey && templateAssetValue
+    }
+
+    if let templateDivAttributesValue = templateDivAttributes {
+      return !templateDivAttributesValue.isEmpty
+    }
+
+    return false
+  }
+
+  func parseVideoElement() -> String? {
+    let sourceUrls = children().compactMap { try? $0.attr(HTMLRawText.Link.source.rawValue) }
+
+    let highDefinitionSourceURL = sourceUrls.first { $0.contains(HTMLRawText.Video.high.rawValue) }
+
+    return highDefinitionSourceURL ?? sourceUrls.first
+  }
+
+  func parseVideoElementThumbnailUrl() -> String? {
+    try? parent()?.attr(HTMLRawText.Image.dataImage.rawValue)
+  }
+
+  func parseImageElement() -> ImageViewElement {
+    var src = ""
+    var caption: String?
+    var href: String?
+
+    if let parent = parent(),
+      parent.tag().getName() == HTMLRawText.Link.anchor.rawValue {
+      href = try? parent.attr(HTMLRawText.Link.link.rawValue)
+    }
+
+    caption = try? attr(HTMLRawText.Image.dataCaption.rawValue)
+
+    let emptyChildren = children().isEmpty()
+    let emptyChildrensChildren = !emptyChildren ? children()[0].children().isEmpty() : true
+
+    if !emptyChildrensChildren {
+      if let updatedSrc = try? children()[0].children()[0].attr(HTMLRawText.Link.source.rawValue) {
+        src = updatedSrc
+      }
+
+      // - if it's a gif collect attribute data-src instead
+      if src.contains(HTMLRawText.Image.gifExtension.rawValue),
+        let updatedSrc = try? children()[0].children()[0].attr(HTMLRawText.Image.dataSource.rawValue) {
+        src = updatedSrc
+      }
+    }
+
+    return ImageViewElement(src: src, href: href, caption: caption)
+  }
+
+  func parseExternalElement() -> ExternalSourceViewElement {
+    let children = children()
+
+    guard !children.isEmpty(),
+      let firstChildElementWithAttributeApplied = try? children[0]
+      .attr(HTMLRawText.Base.width.rawValue, HTMLRawText.Size.hundredPercent.rawValue).text() else {
+      return ExternalSourceViewElement(htmlContent: "")
+    }
+
+    let externalElementValue = ExternalSourceViewElement(htmlContent: firstChildElementWithAttributeApplied)
+
+    return externalElementValue
+  }
+}
