+++ Library/Project Page Campaign HTML Parser/HTMLRawText.swift
@@ -1,0 +1,40 @@
+import Foundation
+
+public enum HTMLRawText {
+  enum Base: String {
+    case htmlClass = "class"
+    case width
+    case div
+  }
+
+  enum List: String {
+    case listItem = "li"
+    case unorderedList = "ul"
+  }
+
+  enum Size: String {
+    case hundredPercent = "100%"
+  }
+
+  enum Link: String {
+    case anchor = "a"
+    case link = "href"
+    case source = "src"
+  }
+
+  enum Image: String {
+    case dataCaption = "data-caption"
+    case dataSource = "data-src"
+    case dataImage = "data-image"
+    case gifExtension = ".gif"
+  }
+
+  enum Video: String {
+    case high
+  }
+
+  enum KSRSpecific: String {
+    case templateAsset = "template asset"
+    case templateOembed = "template oembed"
+  }
+}
