+++ Library/Project Page Campaign HTML Parser/View Elements/ImageViewElement.swift
@@ -1,5 +1,4 @@
 import Foundation
-import UIKit
 
 struct ImageViewElement: ViewElement {
   var src: String
