+++ Library/Project Page Campaign HTML Parser/View Elements/TextComponent.swift
