+++ Library/Project Page Campaign HTML Parser/View Elements/VideoViewElement.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+struct VideoViewElement: ViewElement {
+  let sourceUrl: String
+  let thumbnailUrl: String?
+  let seekPosition: Int64
+}
