--- UnknownViewElement.swift
@@ -1,5 +1,0 @@
-import Foundation
-
-struct UnknownViewElement: ViewElement, Codable {
-  var text: String
-}
